Initial project file
This is a version that I recently converted from a somewhat messy iteration of my code into a more robust State Machine version. It is much easier to debug and modify in this state. I hope to split this code into several different files to make it much more manageable.